@@ -1,25 +1,19 @@
 {
-  "_args": [
-    [
-      "ajv@6.12.2",
-      "/Volumes/data/turbo-site/data"
-    ]
-  ],
-  "_from": "ajv@6.12.2",
-  "_id": "ajv@6.12.2",
+  "_from": "ajv@6.12.6",
+  "_id": "ajv@6.12.6",
   "_inBundle": false,
-  "_integrity": "sha512-k+V+hzjm5q/Mr8ef/1Y9goCmlsK4I6Sm74teeyGvFk1XrOsbsKLjEdrvny42CZ+a8sXbk8KWpY/bDwS+FLL2UQ==",
+  "_integrity": "sha512-j3fVLgvTo527anyYyJOGTYJbG+vnnQYvE0m5mmkc1TK+nxAppkCLMIL0aZ4dblVCNoGShhm+kzE4ZUykBoMg4g==",
   "_location": "/ajv",
   "_phantomChildren": {},
   "_requested": {
     "type": "version",
     "registry": true,
-    "raw": "ajv@6.12.2",
+    "raw": "ajv@6.12.6",
     "name": "ajv",
     "escapedName": "ajv",
-    "rawSpec": "6.12.2",
+    "rawSpec": "6.12.6",
     "saveSpec": null,
-    "fetchSpec": "6.12.2"
+    "fetchSpec": "6.12.6"
   },
   "_requiredBy": [
     "/eslint",
@@ -38,15 +32,17 @@
     "/webpack-dev-server/schema-utils",
     "/webpack/schema-utils"
   ],
-  "_resolved": "https://registry.npmjs.org/ajv/-/ajv-6.12.2.tgz",
-  "_spec": "6.12.2",
-  "_where": "/Volumes/data/turbo-site/data",
+  "_resolved": "https://registry.npmjs.org/ajv/-/ajv-6.12.6.tgz",
+  "_shasum": "baf5a62e802b07d977034586f8c3baf5adf26df4",
+  "_spec": "ajv@6.12.6",
+  "_where": "/Volumes/data/turbo-site/data/node_modules/eslint",
   "author": {
     "name": "Evgeny Poberezkin"
   },
   "bugs": {
-    "url": "https://github.com/epoberezkin/ajv/issues"
+    "url": "https://github.com/ajv-validator/ajv/issues"
   },
+  "bundleDependencies": false,
   "collective": {
     "type": "opencollective",
     "url": "https://opencollective.com/ajv"
@@ -57,6 +53,7 @@
     "json-schema-traverse": "^0.4.1",
     "uri-js": "^4.2.2"
   },
+  "deprecated": false,
   "description": "Another JSON Schema Validator",
   "devDependencies": {
     "ajv-async": "^1.0.0",
@@ -67,7 +64,7 @@
     "coveralls": "^3.0.1",
     "del-cli": "^3.0.0",
     "dot": "^1.0.3",
-    "eslint": "^6.0.0",
+    "eslint": "^7.3.1",
     "gh-pages-generator": "^0.2.3",
     "glob": "^7.0.0",
     "if-node-version": "^1.0.0",
@@ -78,11 +75,11 @@
     "karma-chrome-launcher": "^3.0.0",
     "karma-mocha": "^2.0.0",
     "karma-sauce-launcher": "^4.1.3",
-    "mocha": "^7.0.1",
+    "mocha": "^8.0.1",
     "nyc": "^15.0.0",
     "pre-commit": "^1.1.1",
     "require-globify": "^1.3.0",
-    "typescript": "^2.8.3",
+    "typescript": "^3.9.5",
     "uglify-js": "^3.6.9",
     "watch": "^1.0.0"
   },
@@ -93,7 +90,11 @@
     "LICENSE",
     ".tonic_example.js"
   ],
-  "homepage": "https://github.com/epoberezkin/ajv",
+  "funding": {
+    "type": "github",
+    "url": "https://github.com/sponsors/epoberezkin"
+  },
+  "homepage": "https://github.com/ajv-validator/ajv",
   "keywords": [
     "JSON",
     "schema",
@@ -119,7 +120,7 @@
   },
   "repository": {
     "type": "git",
-    "url": "git+https://github.com/epoberezkin/ajv.git"
+    "url": "git+https://github.com/ajv-validator/ajv.git"
   },
   "scripts": {
     "build": "del-cli lib/dotjs/*.js \"!lib/dotjs/index.js\" && node scripts/compile-dots.js",
@@ -142,5 +143,5 @@
   },
   "tonicExampleFilename": ".tonic_example.js",
   "typings": "lib/ajv.d.ts",
-  "version": "6.12.2"
+  "version": "6.12.6"
 }
