@@ -1,33 +1,27 @@
 {
-  "_args": [
-    [
-      "follow-redirects@1.5.10",
-      "/Volumes/data/turbo-site/data"
-    ]
-  ],
-  "_from": "follow-redirects@1.5.10",
-  "_id": "follow-redirects@1.5.10",
+  "_from": "follow-redirects@1.15.1",
+  "_id": "follow-redirects@1.15.1",
   "_inBundle": false,
-  "_integrity": "sha512-0V5l4Cizzvqt5D44aTXbFZz+FtyXV1vrDN6qrelxtfYQKW0KO0W2T/hkE8xvGa/540LkZlkaUjO4ailYTFtHVQ==",
+  "_integrity": "sha512-yLAMQs+k0b2m7cVxpS1VKJVvoz7SS9Td1zss3XRwXj+ZDH00RJgnuLx7E44wx02kQLrdM3aOOy+FpzS7+8OizA==",
   "_location": "/follow-redirects",
   "_phantomChildren": {},
   "_requested": {
     "type": "version",
     "registry": true,
-    "raw": "follow-redirects@1.5.10",
+    "raw": "follow-redirects@1.15.1",
     "name": "follow-redirects",
     "escapedName": "follow-redirects",
-    "rawSpec": "1.5.10",
+    "rawSpec": "1.15.1",
     "saveSpec": null,
-    "fetchSpec": "1.5.10"
+    "fetchSpec": "1.15.1"
   },
   "_requiredBy": [
-    "/axios",
     "/http-proxy"
   ],
-  "_resolved": "https://registry.npmjs.org/follow-redirects/-/follow-redirects-1.5.10.tgz",
-  "_spec": "1.5.10",
-  "_where": "/Volumes/data/turbo-site/data",
+  "_resolved": "https://registry.npmjs.org/follow-redirects/-/follow-redirects-1.15.1.tgz",
+  "_shasum": "0ca6a452306c9b276e4d3127483e29575e207ad5",
+  "_spec": "follow-redirects@1.15.1",
+  "_where": "/Volumes/data/turbo-site/data/node_modules/http-proxy",
   "author": {
     "name": "Ruben Verborgh",
     "email": "ruben@verborgh.org",
@@ -36,6 +30,7 @@
   "bugs": {
     "url": "https://github.com/follow-redirects/follow-redirects/issues"
   },
+  "bundleDependencies": false,
   "contributors": [
     {
       "name": "Olivier Lalonde",
@@ -47,26 +42,27 @@
       "email": "james@talmage.io"
     }
   ],
-  "dependencies": {
-    "debug": "=3.1.0"
-  },
+  "deprecated": false,
   "description": "HTTP and HTTPS modules that follow redirects.",
   "devDependencies": {
-    "concat-stream": "^1.6.0",
-    "coveralls": "^3.0.2",
-    "eslint": "^4.19.1",
-    "express": "^4.16.2",
-    "mocha": "^5.0.0",
-    "nyc": "^11.8.0"
+    "concat-stream": "^2.0.0",
+    "eslint": "^5.16.0",
+    "express": "^4.16.4",
+    "lolex": "^3.1.0",
+    "mocha": "^6.0.2",
+    "nyc": "^14.1.1"
   },
   "engines": {
     "node": ">=4.0"
   },
   "files": [
-    "index.js",
-    "create.js",
-    "http.js",
-    "https.js"
+    "*.js"
+  ],
+  "funding": [
+    {
+      "type": "individual",
+      "url": "https://github.com/sponsors/RubenVerborgh"
+    }
   ],
   "homepage": "https://github.com/follow-redirects/follow-redirects",
   "keywords": [
@@ -81,11 +77,10 @@
   "license": "MIT",
   "main": "index.js",
   "name": "follow-redirects",
-  "nyc": {
-    "reporter": [
-      "lcov",
-      "text"
-    ]
+  "peerDependenciesMeta": {
+    "debug": {
+      "optional": true
+    }
   },
   "repository": {
     "type": "git",
@@ -96,5 +91,5 @@
     "mocha": "nyc mocha",
     "test": "npm run lint && npm run mocha"
   },
-  "version": "1.5.10"
+  "version": "1.15.1"
 }
