@@ -1,32 +1,27 @@
 {
-  "_args": [
-    [
-      "strip-ansi@4.0.0",
-      "/Volumes/data/turbo-site/data"
-    ]
-  ],
-  "_from": "strip-ansi@4.0.0",
-  "_id": "strip-ansi@4.0.0",
+  "_from": "strip-ansi@^6.0.1",
+  "_id": "strip-ansi@6.0.1",
   "_inBundle": false,
-  "_integrity": "sha1-qEeQIusaw2iocTibY1JixQXuNo8=",
+  "_integrity": "sha512-Y38VPSHcqkFrCpFnQ9vuSXmquuv5oXOKpGeT6aGrr3o3Gc9AlVa6JBfUSOCnbxGGZF+/0ooI7KrPuUSztUdU5A==",
   "_location": "/string-width/strip-ansi",
   "_phantomChildren": {},
   "_requested": {
-    "type": "version",
+    "type": "range",
     "registry": true,
-    "raw": "strip-ansi@4.0.0",
+    "raw": "strip-ansi@^6.0.1",
     "name": "strip-ansi",
     "escapedName": "strip-ansi",
-    "rawSpec": "4.0.0",
+    "rawSpec": "^6.0.1",
     "saveSpec": null,
-    "fetchSpec": "4.0.0"
+    "fetchSpec": "^6.0.1"
   },
   "_requiredBy": [
     "/string-width"
   ],
-  "_resolved": "https://registry.npmjs.org/strip-ansi/-/strip-ansi-4.0.0.tgz",
-  "_spec": "4.0.0",
-  "_where": "/Volumes/data/turbo-site/data",
+  "_resolved": "https://registry.npmjs.org/strip-ansi/-/strip-ansi-6.0.1.tgz",
+  "_shasum": "9e26c63d30f53443e9489495b2105d37b67a85d9",
+  "_spec": "strip-ansi@^6.0.1",
+  "_where": "/Volumes/data/turbo-site/data/node_modules/string-width",
   "author": {
     "name": "Sindre Sorhus",
     "email": "sindresorhus@gmail.com",
@@ -35,19 +30,23 @@
   "bugs": {
     "url": "https://github.com/chalk/strip-ansi/issues"
   },
+  "bundleDependencies": false,
   "dependencies": {
-    "ansi-regex": "^3.0.0"
+    "ansi-regex": "^5.0.1"
   },
-  "description": "Strip ANSI escape codes",
+  "deprecated": false,
+  "description": "Strip ANSI escape codes from a string",
   "devDependencies": {
-    "ava": "*",
-    "xo": "*"
+    "ava": "^2.4.0",
+    "tsd": "^0.10.0",
+    "xo": "^0.25.3"
   },
   "engines": {
-    "node": ">=4"
+    "node": ">=8"
   },
   "files": [
-    "index.js"
+    "index.js",
+    "index.d.ts"
   ],
   "homepage": "https://github.com/chalk/strip-ansi#readme",
   "keywords": [
@@ -81,7 +80,7 @@
     "url": "git+https://github.com/chalk/strip-ansi.git"
   },
   "scripts": {
-    "test": "xo && ava"
+    "test": "xo && ava && tsd"
   },
-  "version": "4.0.0"
+  "version": "6.0.1"
 }
