@@ -1,37 +1,27 @@
 {
-  "_args": [
-    [
-      "string-width@2.1.1",
-      "/Volumes/data/turbo-site/data"
-    ]
-  ],
-  "_from": "string-width@2.1.1",
-  "_id": "string-width@2.1.1",
+  "_from": "string-width@4.2.3",
+  "_id": "string-width@4.2.3",
   "_inBundle": false,
-  "_integrity": "sha512-nOqH59deCq9SRHlxq1Aw85Jnt4w6KvLKqWVik6oA9ZklXLNIOlqg4F2yrT1MVaTjAqvVwdfeZ7w7aCvJD7ugkw==",
+  "_integrity": "sha512-wKyQRQpjJ0sIp62ErSZdGsjMJWsap5oRNihHhu6G7JVO/9jIB6UyevL+tXuOqrng8j/cxKTWyWUwvSTriiZz/g==",
   "_location": "/string-width",
-  "_phantomChildren": {
-    "ansi-regex": "3.0.0"
-  },
+  "_phantomChildren": {},
   "_requested": {
     "type": "version",
     "registry": true,
-    "raw": "string-width@2.1.1",
+    "raw": "string-width@4.2.3",
     "name": "string-width",
     "escapedName": "string-width",
-    "rawSpec": "2.1.1",
+    "rawSpec": "4.2.3",
     "saveSpec": null,
-    "fetchSpec": "2.1.1"
+    "fetchSpec": "4.2.3"
   },
   "_requiredBy": [
-    "/@pieh/friendly-errors-webpack-plugin",
-    "/inquirer",
-    "/react-dev-utils/inquirer",
     "/wide-align"
   ],
-  "_resolved": "https://registry.npmjs.org/string-width/-/string-width-2.1.1.tgz",
-  "_spec": "2.1.1",
-  "_where": "/Volumes/data/turbo-site/data",
+  "_resolved": "https://registry.npmjs.org/string-width/-/string-width-4.2.3.tgz",
+  "_shasum": "269c7117d27b05ad2e536830a8ec895ef9c6d010",
+  "_spec": "string-width@4.2.3",
+  "_where": "/Volumes/data/turbo-site/data/node_modules/wide-align",
   "author": {
     "name": "Sindre Sorhus",
     "email": "sindresorhus@gmail.com",
@@ -40,27 +30,30 @@
   "bugs": {
     "url": "https://github.com/sindresorhus/string-width/issues"
   },
+  "bundleDependencies": false,
   "dependencies": {
-    "is-fullwidth-code-point": "^2.0.0",
-    "strip-ansi": "^4.0.0"
+    "emoji-regex": "^8.0.0",
+    "is-fullwidth-code-point": "^3.0.0",
+    "strip-ansi": "^6.0.1"
   },
+  "deprecated": false,
   "description": "Get the visual width of a string - the number of columns required to display it",
   "devDependencies": {
-    "ava": "*",
-    "xo": "*"
+    "ava": "^1.4.1",
+    "tsd": "^0.7.1",
+    "xo": "^0.24.0"
   },
   "engines": {
-    "node": ">=4"
+    "node": ">=8"
   },
   "files": [
-    "index.js"
+    "index.js",
+    "index.d.ts"
   ],
   "homepage": "https://github.com/sindresorhus/string-width#readme",
   "keywords": [
     "string",
-    "str",
     "character",
-    "char",
     "unicode",
     "width",
     "visual",
@@ -89,7 +82,7 @@
     "url": "git+https://github.com/sindresorhus/string-width.git"
   },
   "scripts": {
-    "test": "xo && ava"
+    "test": "xo && ava && tsd"
   },
-  "version": "2.1.1"
+  "version": "4.2.3"
 }
