@@ -1,9 +1,9 @@
 var plugins = [{
-      plugin: require('/Volumes/data/turbo-site/data/node_modules/gatsby-plugin-google-fonts/gatsby-ssr'),
-      options: {"plugins":[],"fonts":["Teko","Poppins:300,400,500,600,700"],"display":"swap"},
-    },{
       plugin: require('/Volumes/data/turbo-site/data/node_modules/gatsby-plugin-styled-components/gatsby-ssr'),
       options: {"plugins":[],"minify":false,"transpileTemplateLiterals":false,"pure":false,"ssr":true},
+    },{
+      plugin: require('/Volumes/data/turbo-site/data/node_modules/gatsby-plugin-intl/gatsby-ssr'),
+      options: {"plugins":[],"path":"/Volumes/data/turbo-site/data/src/intl","languages":["ro","ru","en"],"defaultLanguage":"ro","redirect":true,"redirectComponent":"/Volumes/data/turbo-site/data/src/components/redirect/index.js"},
     },{
       plugin: require('/Volumes/data/turbo-site/data/node_modules/gatsby-plugin-react-helmet/gatsby-ssr'),
       options: {"plugins":[]},
