@@ -2,12 +2,8 @@
 const preferDefault = m => m && m.default || m
 
 exports.components = {
+  "component---cache-dev-404-page-js": () => import("./dev-404-page.js" /* webpackChunkName: "component---cache-dev-404-page-js" */),
   "component---src-pages-404-js": () => import("./../src/pages/404.js" /* webpackChunkName: "component---src-pages-404-js" */),
-  "component---src-pages-creative-2-index-js": () => import("./../src/pages/creative-2/index.js" /* webpackChunkName: "component---src-pages-creative-2-index-js" */),
-  "component---src-pages-creative-video-1-index-js": () => import("./../src/pages/creative-video-1/index.js" /* webpackChunkName: "component---src-pages-creative-video-1-index-js" */),
-  "component---src-pages-creative-video-2-index-js": () => import("./../src/pages/creative-video-2/index.js" /* webpackChunkName: "component---src-pages-creative-video-2-index-js" */),
-  "component---src-pages-index-js": () => import("./../src/pages/index.js" /* webpackChunkName: "component---src-pages-index-js" */),
-  "component---src-pages-personal-index-js": () => import("./../src/pages/personal/index.js" /* webpackChunkName: "component---src-pages-personal-index-js" */),
-  "component---src-pages-personal-video-index-js": () => import("./../src/pages/personal-video/index.js" /* webpackChunkName: "component---src-pages-personal-video-index-js" */)
+  "component---src-pages-index-js": () => import("./../src/pages/index.js" /* webpackChunkName: "component---src-pages-index-js" */)
 }
 
