@@ -1,14 +1,17 @@
-import React from 'react'
-import './styles.scss'
+import React from "react";
+import "./styles.scss";
 //Styeld Component does not support @for loops
+import { useIntl } from "gatsby-plugin-intl";
 
-class Glitch extends React.Component {
+const Glitch = (props) => {
 
-    render() {
-        const { text } = this.props
-        return(
-            <p className="glitch" data-text={text}>{text}</p>
-        )
-    }
-}
-export default Glitch
+  const intl = useIntl();
+  const text = intl.formatMessage({'id': `${ props.text }`})
+
+  return (
+    <p className="glitch" data-text={text}>
+      {text}
+    </p>
+  );
+};
+export default Glitch;
