@@ -2,7 +2,7 @@ import React from 'react'
 export default class HTML extends React.Component {
   render() {
     return (
-      <html lang="en">
+      <html lang="enz">
         <head>
           <meta charSet="utf-8" />
           <meta
