@@ -1,55 +1,47 @@
-import React from 'react'
-import Layout from 'components/layout'
-import { graphql } from 'gatsby'
-import Helmet from 'react-helmet'
-import Hero from 'sections/hero/Hero.js'
-import AboutOne from 'sections/about/AboutOne.js'
-import ServicesOne from 'sections/services/ServicesOne.js'
-import PortfolioOne from 'sections/portfolio/PortfolioOne.js'
-import TestimonialsOne from 'sections/testimonials/TestimonialsOne.js'
-import TeamOne from 'sections/team/TeamOne.js'
-import ClientsOne from 'sections/clients/ClientsOne.js'
-import ContactCreative from 'sections/contact/ContactCreative.js'
+import React from "react";
+import Layout from "components/layout";
+import HeroPersonal from "sections/hero/HeroPersonal.js";
+import AboutTwo from "sections/about/AboutTwo.js";
+import ServicesOne from "sections/services/ServicesTwo.js";
+import PortfolioThree from "sections/portfolio/PortfolioThree.js";
+import TestimonialsOne from "sections/testimonials/TestimonialsOne.js";
+import ContactCreative from "sections/contact/ContactCreative.js";
+import SEO from "../components/seo";
+import { useIntl } from "gatsby-plugin-intl";
 
-class Index extends React.Component {
+const Index = (props) => {
+  const intl = useIntl();
 
-  render() {
+  return (
+    <div>
+      <Layout
+        isHome={true}
+        sections={[
+          "home",
+          "about",
+          "services",
+          "portfolio",
+          "testimonials",
+          "contact",
+        ]}
+      >
+        <SEO
+          lang={intl.locale}
+          title={intl.formatMessage({ id: "title" })}
+          keywords={[`разработка`, `сайтой`, `Молдова`, `Кишинев`, `интернет-магазин`, `создать`]}
+          description={intl.formatMessage({ id: "description" })}
+        />
 
-    const { site } = this.props.data
+        <HeroPersonal />
+        <AboutTwo />
+        <ServicesOne />
+        <PortfolioThree />
+        <TestimonialsOne />
+        <ContactCreative />
+      </Layout>
+    </div>
+  );
+};
 
-    return (
-      <div>
-        <Helmet>
-          <title>{site.meta.title}</title>
-          <meta name="description" content={site.meta.description} />
-        </Helmet>
-        <Layout
-          isHome={true}
-          sections={['home', 'about', 'services', 'portfolio', 'testimonials', 'team', 'clients', 'contact']}
-        >
-          <Hero />
-          <AboutOne />
-          <ServicesOne />
-          <PortfolioOne />
-          <TestimonialsOne />
-          <TeamOne />
-          <ClientsOne />
-          <ContactCreative />
-        </Layout>
-      </div>
-    )
-  }
-}
+export default Index;
 
-export default Index
-
-export const creativeOneQuery = graphql`
-  query {
-    site {
-      meta: siteMetadata {
-        title
-        description
-      }
-    }
-  }
-`
