@@ -3,7 +3,7 @@ import { Row, Col, Container } from 'react-bootstrap'
 import { StaticQuery, graphql } from 'gatsby'
 import styled from 'styled-components'
 import Typewriter from 'typewriter-effect'
-import LoopVideo from './assets/loop.mp4'
+import LoopVideo from './assets/home6.mp4'
 
 class HeroVideo extends React.Component {
 
@@ -46,7 +46,6 @@ class HeroVideo extends React.Component {
             background-size: cover;
             background-repeat: no-repeat;
         `
-
         const VideoContainer = styled.div`
             width: 100%;
             display: flex;
@@ -56,10 +55,20 @@ class HeroVideo extends React.Component {
             video {
                 position: absolute;
             }
+            video:before{
+                content: '';
+                display: block;
+                position: absolute;
+                top:0;
+                left:0;
+                height: 100%;
+                width:100%;
+                background-color: rgba(0,0,0,.5);
+            }
         `
         const Heading1 = styled.h1`
             font-size: 80px;
-            font-family: Teko;
+            font-family: Oswald;
             line-height: 40px;
             color: #fff;
             font-weight: 400;
@@ -119,12 +128,16 @@ class HeroVideo extends React.Component {
                 line-height: 20px;
             }
             span {
-                font-family: Teko;
+                font-family: Oswald;
             }
         `
         const Content = styled.div`
             position: relative;
-            z-index: 10
+            z-index: 10;  
+            padding:35px 10px;
+            background: rgb(255,255,255,.0);
+            border-radius:1px;
+            margin: 0 -15px;
         `
 
         const Separator = styled.div`
@@ -157,66 +170,12 @@ class HeroVideo extends React.Component {
             }
         `
 
-        const ServiceCol = styled(Col)`
-            padding: 0;
-            border-right: #444;
-            &.no-border {
-                border-right: none;
-            }
-            @media (max-width:500px) {
-                border-right:none;
-            }
-        `
 
-        const Service = styled.div`
-            min-height: 300px;
-            background-color: #111;
-            transition: .2s;
-            display: flex;
-            justify-content: center;
-            flex-direction: column;
-            &:hover {
-                background-color: #222;
-            }
-        `
-        const ServiceContent = styled.div`
-            padding: 40px;
-            color: #fff;
-            p {
-                font-size: 14px;
-                font-weight: 300;
-                color: #efefef;
-            }
-        `
-        const ServiceHeading = styled.h4`
-            font-size: 30px;
-            font-weight: 400;
-            font-family: Teko;
-        `
-        const ServiceSeparator = styled.div`
-            height: 5px;
-            width: 50px;
-            background-color: #04e5e5;
-            margin-bottom: 10px;
-        `
-
-        const ServiceIcon = styled.div`
-            margin-bottom: 20px;
-            img {
-                max-height: 70px;
-            }
-        `
-
-        const ServiceText = styled.p`
-            font-size: 14px;
-            font-weight: 300;
-            color: #c5c5c5;
-        `
 
         return (
-            <Section id="home">
+            <Section id="home"> 
                 <VideoContainer style={{height: `${this.state.width > 500 ? this.state.height : 350}px`}}>
-                    <video autoPlay="autoplay" loop="loop" muted style={{height: `${this.state.width >= 1024 && this.state.width < 1200 ? "100%": "auto"}`}}>
+                    <video className="home-video" autoPlay="autoplay" loop="loop" muted style={{height: `${this.state.width >= 1024 && this.state.width < 1200 ? "100%": "auto"}`, maxWidth: '100%' }}>
                         <source src={LoopVideo} type="video/mp4" />
                     </video>
                     <Container>
@@ -246,53 +205,7 @@ class HeroVideo extends React.Component {
                 </VideoContainer>
                 <BottomContent>
                         <Container>
-                            <Row>
-                                <ServiceCol md={4} style={{borderRight: this.state.width > 500 ? "1px solid #444" : "none"}}>
-                                    <Service>
-                                        <ServiceContent>
-                                            <ServiceIcon>
-                                                <img src={this.props.webexpertIcon.childImageSharp.fluid.src} alt="web experts" />
-                                            </ServiceIcon>
-                                            <ServiceHeading>Web Experts</ServiceHeading>
-                                            <ServiceSeparator/>
-                                            <ServiceText>
-                                                Lorem ipsum dolor sit amet, consectetur adipiscing elit. In vitae ultricies lacus, vitae varius velit. 
-                                                Pellentesque blandit libero eu turpis condimentum bibendum.
-                                            </ServiceText>
-                                        </ServiceContent>
-                                    </Service>
-                                </ServiceCol>
-                                <ServiceCol md={4} style={{borderRight: this.state.width > 500 ? "1px solid #444" : "none"}}>
-                                    <Service>
-                                        <ServiceContent>
-                                            <ServiceIcon>
-                                                    <img src={this.props.mobileIcon.childImageSharp.fluid.src} alt="mobile experts" />
-                                            </ServiceIcon>
-                                            <ServiceHeading>Mobile Experts</ServiceHeading>
-                                            <ServiceSeparator/>
-                                            <ServiceText>
-                                                Lorem ipsum dolor sit amet, consectetur adipiscing elit. In vitae ultricies lacus, vitae varius velit. 
-                                                Pellentesque blandit libero eu turpis condimentum bibendum.
-                                            </ServiceText>
-                                        </ServiceContent>
-                                    </Service>
-                                </ServiceCol>
-                                <ServiceCol md={4}>
-                                    <Service>
-                                        <ServiceContent>
-                                            <ServiceIcon>
-                                                    <img src={this.props.seoIcon.childImageSharp.fluid.src} alt="seo experts" />
-                                            </ServiceIcon>
-                                            <ServiceHeading>SEO Experts</ServiceHeading>
-                                            <ServiceSeparator/>
-                                            <ServiceText>
-                                                Lorem ipsum dolor sit amet, consectetur adipiscing elit. In vitae ultricies lacus, vitae varius velit. 
-                                                Pellentesque blandit libero eu turpis condimentum bibendum.
-                                            </ServiceText>
-                                        </ServiceContent>
-                                    </Service>
-                                </ServiceCol>
-                            </Row>
+             
                         </Container>
                 </BottomContent>
             </Section>
@@ -311,7 +224,7 @@ export default props => (
             }
           }
         }
-        webexpertIcon: file(relativePath: {eq: "icons/web.png"}) {
+        webexpertIcon: file(relativePath: {eq: "icons/shop.png"}) {
           childImageSharp {
             fluid(maxWidth: 500) {
               src
@@ -325,7 +238,7 @@ export default props => (
             }
           }
         }
-        seoIcon: file(relativePath: {eq: "icons/seo.png"}) {
+        seoIcon: file(relativePath: {eq: "icons/landing.png"}) {
           childImageSharp {
             fluid(maxWidth: 500) {
               src
